AdminLogin -> AdminLoginPage and moved it to pages folder
